+++ packages/@react-elf-types/ghost/src/index.d.ts
@@ -1,0 +1,8 @@
+import type { ReactNode } from "react";
+import { CommonStyle } from "@react-elf-types/shared";
+
+export interface GhostProps {
+  animated?: boolean;
+  style?: CommonStyle;
+  children?: ReactNode;
+}