+++ packages/@react-elf-types/layout/src/index.d.ts
@@ -1,8 +1,8 @@
 import { CommonStyle } from "@react-elf-types/shared";
 
 export type GridStyle = {
-  backgroundColor: string;
-  gap: number;
+  backgroundColor?: string;
+  gap?: number;
 };
 
 export type GridTemplate = number | string;
@@ -27,7 +27,6 @@ export interface ColumnProps {
   style?: CommonStyle;
 }
 
-
 export type FlexStyle = {
   backgroundColor?: string;
   gap?: number;
@@ -40,10 +39,24 @@ export interface FlexProps {
   wrap?: boolean;
   sameWidth?: boolean;
   style?: FlexStyle & CommonStyle;
+  direction?: "row" | "column";
+  gap?: number | string;
+  columnGap?: number | string;
+  rowGap?: number | string;
+  justifyContent?:
+    | "start"
+    | "center"
+    | "end"
+    | "flex-start"
+    | "flex-end"
+    | "space-around"
+    | "space-evenly"
+    | "space-between";
 }
 
 export interface VBoxProps {
   children?: React.ReactNode;
   wrap?: boolean;
   style?: FlexStyle & CommonStyle;
+  gap?: number | string;
 }