+++ packages/@react-elf/layout/src/flex/VBox.tsx
@@ -3,10 +3,10 @@ import { VBoxProps } from "@react-elf-types/layout";
 import { Flex } from "./Flex";
 
 export function VBox(props: VBoxProps) {
-  const { style = {}, children } = props;
+  const { style = {}, children, gap = undefined } = props;
 
   return (
-    <Flex stack={true} style={style}>
+    <Flex stack style={style} gap={gap}>
       {children}
     </Flex>
   );