+++ packages/@react-elf/ui/src/index.ts
@@ -12,3 +12,4 @@ export * from "@react-elf/toast";
 export * from "@react-elf/dialog";
 export * from "@react-elf/toolbar";
 export * from "@react-elf/tools";
+export * from "@react-elf/ghost";