+++ packages/@react-elf-types/divider/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @react-elf-types/divider
 
+## 0.0.66
+
+### Patch Changes
+
+- add ghost, divider component
+
+- Updated dependencies []:
+  - @react-elf-types/shared@0.0.66
+
 ## 0.0.65
 
 ### Patch Changes