+++ packages/@react-elf-types/shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @react-elf-types/shared
 
+## 0.0.66
+
+### Patch Changes
+
+- add ghost, divider component
+
 ## 0.0.65
 
 ### Patch Changes