+++ packages/@react-elf/button/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @react-elf/button
 
+## 0.0.66
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @react-elf-types/button@0.0.66
+  - @react-elf/shared@0.0.66
+  - @react-elf/tooltip@0.0.66
+
 ## 0.0.65
 
 ### Patch Changes