+++ packages/@react-elf/layout/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @react-elf/layout
 
+## 0.0.66
+
+### Patch Changes
+
+- add ghost, divider component
+
+- Updated dependencies []:
+  - @react-elf-types/layout@0.0.66
+  - @react-elf/shared@0.0.66
+
 ## 0.0.65
 
 ### Patch Changes