+++ packages/@react-elf/shared/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @react-elf/button
 
+## 0.0.66
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @react-elf-types/shared@0.0.66
+
 ## 0.0.65
 
 ### Patch Changes