+++ packages/@react-elf/toolbar/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @react-elf/toolbar
 
+## 0.0.66
+
+### Patch Changes
+
+- add ghost, divider component
+
+- Updated dependencies []:
+  - @react-elf-types/shared@0.0.66
+  - @react-elf-types/toolbar@0.0.66
+  - @react-elf/shared@0.0.66
+  - @react-elf/tools@0.0.66
+
 ## 0.0.65
 
 ### Patch Changes