+++ packages/@react-elf/tools/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @react-elf/tools
 
+## 0.0.66
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @react-elf/layout@0.0.66
+  - @react-elf-types/tools@0.0.66
+  - @react-elf-types/tooltip@0.0.66
+  - @react-elf/shared@0.0.66
+  - @react-elf/menu@0.0.66
+  - @react-elf/tooltip@0.0.66
+
 ## 0.0.65
 
 ### Patch Changes